+++ hexl/experimental/fft/fft.cpp
@@ -1,9 +1,9 @@
 // Copyright (C) 2020-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "hexl/fft/fft.hpp"
+#include "hexl/experimental/fft/fft.hpp"
 
-#include "hexl/fft/fft-native.hpp"
+#include "hexl/experimental/fft/fft-native.hpp"
 #include "hexl/logging/logging.hpp"
 
 namespace intel {
