+++ hexl/include/hexl/experimental/fft/fft.hpp
@@ -5,8 +5,8 @@
 
 #include <complex>
 
-#include "hexl/fft/fwd-fft-avx512.hpp"
-#include "hexl/fft/inv-fft-avx512.hpp"
+#include "hexl/experimental/fft/fwd-fft-avx512.hpp"
+#include "hexl/experimental/fft/inv-fft-avx512.hpp"
 #include "hexl/util/aligned-allocator.hpp"
 #include "hexl/util/allocator.hpp"
