+++ test/CMakeLists.txt
@@ -12,8 +12,6 @@ set(NATIVE_TEST_SRC main.cpp
     test-eltwise-reduce-mod.cpp
     test-eltwise-sub-mod.cpp
     test-ntt.cpp
-    test-fft.cpp
-    test-fft-native.cpp
     test-util-internal.cpp
 )
 
@@ -22,6 +20,9 @@ if (HEXL_EXPERIMENTAL)
         experimental/seal/test-dyadic-multiply.cpp
         experimental/seal/test-key-switch.cpp
         experimental/misc/test-lr-mat-vec-mult.cpp
+        experimental/fft/test-fft-avx512.cpp
+        experimental/fft/test-fft.cpp
+        experimental/fft/test-fft-native.cpp
     )
 endif()
 
@@ -35,7 +36,6 @@ set(AVX512_TEST_SRC
     test-eltwise-reduce-mod-avx512.cpp
     test-eltwise-sub-mod-avx512.cpp
     test-ntt-avx512.cpp
-    test-fft-avx512.cpp
 )
 
 set(TEST_SRC "${NATIVE_TEST_SRC};${AVX512_TEST_SRC}")
