+++ test/experimental/fft/test-fft-avx512.cpp
@@ -5,8 +5,8 @@
 
 #include <complex>
 
-#include "hexl/fft/fft-avx512-util.hpp"
-#include "hexl/fft/fft.hpp"
+#include "hexl/experimental/fft/fft-avx512-util.hpp"
+#include "hexl/experimental/fft/fft.hpp"
 #include "hexl/logging/logging.hpp"
 #include "hexl/util/defines.hpp"
 #include "test-util.hpp"
