+++ test/test-ntt-avx512.cpp
@@ -6,7 +6,7 @@
 #include <tuple>
 #include <vector>
 
-#include "hexl/fft/fft.hpp"
+#include "hexl/experimental/fft/fft.hpp"
 #include "hexl/ntt/ntt.hpp"
 #include "hexl/number-theory/number-theory.hpp"
 #include "ntt/fwd-ntt-avx512.hpp"
